+++ GAME3023 - RPG/Assets/Scripts/Battle/UI/BattleSceneUI.cs
@@ -108,6 +108,6 @@ public class BattleSceneUI : MonoBehaviour
     {
         // add probability stuff later
 
-        GameManager.instance.sceneLoader.LoadNextScene(SceneDirectory.Battle,SceneDirectory.PlainsCenter);
+        GameManager.instance.sceneLoader.LoadNextScene((int)SceneDirectory.Battle,SceneDirectory.PlainsCenter);
     }
 }
